feat: Fix USD integration (#444)
* fix: Remove unused components

* fix: fix tests

* fix: Fix typescript errors

* fix: correctly load stories
- remove unused stories
- readd useful stories
