--- app/screens/receive-bitcoin-screen/receiv-bitcoin-screen.types.d.ts
@@ -1,8 +1,0 @@
-type GetFullUriInput = {
-  input: string
-  amount?: number
-  memo?: string
-  uppercase?: boolean
-  prefix?: boolean
-  type?: "BITCOIN_ONCHAIN" | "LIGHTNING_BTC" | "LIGHTNING_USD"
-}
