+++ app/screens/receive-bitcoin-screen/receive-bitcoin-screen.types.d.ts
@@ -1,5 +1,8 @@
 type GetFullUriInput = {
   input: string
+  amount?: number
+  memo?: string
   uppercase?: boolean
   prefix?: boolean
+  type?: "BITCOIN_ONCHAIN" | "LIGHTNING_BTC" | "LIGHTNING_USD"
 }
