+++ app/screens/send-bitcoin-screen/index.ts
@@ -1,4 +1,2 @@
 export * from "./use-fee"
-export * from "./send-bitcoin-screen"
 export * from "./scanning-qrcode-screen"
-export * from "./send-bitcoin-confirmation-screen"
