+++ app/screens/send-bitcoin-screen/send-bitcoin-amount.tsx
@@ -499,7 +499,7 @@ const SendBitcoinAmount = ({
           </View>
         </View>
       </View>
-      <View style={Styles.buttonContainer}>
+      <View >
         <Button
           title={
             !validate()
