+++ storybook/storybook-registry.ts
@@ -1,20 +1,7 @@
-require("../app/components/balance-header/balance-header.story")
 require("../app/components/input-payment/input-payment.story")
 require("../app/components/price-graph/price-graph.story")
 require("../app/components/large-button/large-button.story")
-
-// require("../app/screens/earns-screen/earns-quiz.story")
-// require("../app/screens/earns-screen/section-completed.story")
-// require("../app/screens/earns-map-screen/earns-map-screen.story")
 require("../app/screens/transaction-detail-screen/transaction-detail-screen.story")
-// require("../app/screens/transaction-screen/transaction-screen.story")
-// require("../app/screens/move-money-screen/move-money-screen.story")
-// require("../app/screens/settings-screen/settings-screen.story")
-// require("../app/screens/settings-screen/username-screen.story")
-// require("../app/screens/send-bitcoin-screen/send-bitcoin-screen.story")
-// require("../app/screens/phone-auth-screen/phone-auth.story")
-// require("../app/screens/contacts-screen/contacts.story")
-// require("../app/screens/contacts-screen/contacts-detail.story")
 require("../app/components/wallet-overview/wallet-overview.story")
 require("../app/components/wallet-summary/wallet-summary.story")
